voice-service/requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[daily,openai,google,silero,deepgram]
google-generativeai
